requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
mysql-connector-python
pickle
# json

# notebook
python-box==6.0.2
pyYAML
ensure==1.0.2
joblib
types-PyYAML
gradio

-e .
